requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
requests==2.31.0
peewee==3.16.2
python-dateutil==2.8.2
loguru==0.7.2
